move State under Ciri module
